Move ProfileContextCheck to domain
