+++ backend/src/main/java/com/alligator/market/backend/provider/reconciliation/scanner/ProfileContextScannerAdapter.java
@@ -1,6 +1,7 @@
 package com.alligator.market.backend.provider.reconciliation.scanner;
 
 import com.alligator.market.domain.provider.contract.MarketDataProvider;
+import com.alligator.market.domain.provider.reconciliation.ProfileContextCheck;
 import com.alligator.market.domain.provider.reconciliation.ProfileContextScanner;
 import com.alligator.market.domain.provider.profile.model.Profile;
 import lombok.RequiredArgsConstructor;
